Cleaned up layout by breaking into partials
Rails allows the creation of multiple layout files --
For example, you could have one layout render for normal users, and a
different layout for admins.

Because there would be a lot of shared code between these (most likely,
we'd want to keep the nav bar and head content in both layouts),
components of the layout should be broken into seperate files
where possible.

In this commit, I added partials under the 'app/views/application'
directory for flash messages, the footer, the head tag, and the nav bar.

The resulting application layout file is much easier to understand, and
now simply deals with the flow and position of the componenets
on the page, instead of the content of the components.