Remove the ability to nest comments
Nesting really isn't necessary for an application like this. In fact,
I'd say that we don't want it.

Instead of allowing users to start up conversations (and arguments)
between each other, having non-nested comments will encourage them to
focus on the project that they're commenting on, instead of other users'
comments.

Plus, a single thread of comments fits in much better with our new
design of the project page, where they'll be displayed in the sidebar
next to the project description.